imap: Move to a single "fetcher" go routine
In preparation for reorganizing the control flow, make a single
"fetcher" goroutine that accepts fetch requests and responds
appropriately.

NB that in spite of an earlier commit, we don't add a mutex.  This is
because we *also* need to make sure that (say) sequence numbers don't
change underneath us.  Since we have to be careful about usage anyway,
just be careful.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>