Actually remove mutex, first steps towards a full pipeline
Apparently removal of mutex didn't make it in to the previous
changeset; include them here because we're lazy.

First step towards a more flow-based processing: Make envelope fetch
requests in batches of 50 from a separate goroutine.

Work around the fact that client.Fetch() closes the message channel by
creating goroutines to read from the channel and pass them on to
another channel, which is not closed.

Count the number of open routines and close the final channel once
they've all completed.

Current setup is suboptimal because the "results" channel must be big
enough to have all results.  This should be reworked somehow in the
future.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>