imap: Make fully pipelined
Pipeline the rest of the fetching process.  Create a goroutine for
every message to fetch; at that point there are no more "forks" in the
processing.  Use the same trick of passing status channels along a
channel to the main channel.

Also note that there's an inherent circularity once you get several loops
adding things to the fetch request queue; this means that we have to prevent
individual requests from blocking.  Make sure that the channels passed to
the fetch goroutine are big enough to hold all the message results.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>