Sync mailbox status
In preparation for using the status of a mail still being in a mailbox
(or not), add the "mailbox" concept.

Add a lmdb_mailboxes table to map names to IDs (so that names can
change), and an lmdb_mailbox_join table to map mailboxes to message
Ids.

Rework ImapInfo to MailboxInfo.  Add a place to stash the mailboxId.
Read it on start-up if it's there, otherwise create it.

We have to collect all message IDs in one list so that we know which
ones have disappeared; collect these Ids, then simply throw away all
the old IDs and re-insert the new ones.

Restructuring is also meant to look forward to when we may have more
than one mailbox to sync.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>