First cut at fetching mail from the database
Create a MessageTree structure for storing a tree.

Add a query to find the 'root' mails of all mails in the inbox.  Do
this by getting the "closure" of all `inreplyto` messageids in the
database, and then finding the ones with no parents in the database
(either because no parent is given, or because the parent is not in
the database).

For now, assume that mailboxId is set, so we don't have to mess around
with transactions.

Also add a hack to mailfetch to call the new function; in the future
we'll make this a proper command-base (probably using cobra or
somtehing).

Signed-off-by: George Dunlap <george.dunlap@citrix.com>