localmaildb: Handle unusual mail formatting
Unfortunately "net/mail" (and also "emersion/go-message/mail") are too
strict for some of the mails in the xen-devel archive: Neither handles
some of the character sets encouuntered (specifically, iso-8859-2);
the former refuses to process mail headers that contain such
characters at all, and the latter refuses to process the entire mail.

For now, make it possible to skip messages that have parse errors.

Add some types that make it possible to call errors.Is()  but still
have richer errors.  Implement this fro ErrorMsgidPresent as well, and
include the offending messageid in the error message.  (Later we
should include this in an extract-able field of a struct.)

Modify pubinbox.go to continue processing if the error returned is an
ErrParseError, but report the total skipped.

Include at least one offending mail for future reference.  One way to
fix this would be to implement a mime.WordDecoder using
x/text/encoding/charmap.

Signed-off-by: George Dunlap <george.dunlap@citrix.com>