imapsource, mailfetch: Port to new localmaildb package setup
Fix imapsource's package name.

Replace 'mdb' with 'src' as an object name for clarity.

s/MailDb/lmdb.MailDB/g;

Read the body of the message using ReadAll and pass it in to
AddMessage() (rather than expecting AddMessage to read it itself).

Add a `Setup` function which takes a MailboxInfo and returns an ImapSource.

Have ImapConnect() use MailboxName for the folder if non-empty
(falling back to "INBOX" if it is).  Also pass this to
UpdateMailbox().

Have mdb.CreateMailbox() ignore an error if the mailbox exists.

Add imapsrc package to mailfetch; call Setup; and switch to fetching
mail again (rather than whatever checking thing we were doing).  Have
it also unconditionally call CreateMailbox().

While here:

- Add some logging & clarify some log messages
- Check the return value to UpdateMailbox so failures aren't silently ignored

Signed-off-by: George Dunlap <george.dunlap@cloud.com>