cli: Add a way to extract an am from a messageid
Add `export-am` command, which will search for a given messageid and
output an mbox with the appropriate messages in it to use as a git-am.

Enumerate the various cases we want to cover, but to begin with
hard-code only the "0/N" pattern (since that was the pattern of the
series we wanted to review last week).

In the course of this, it turned out that the GetTree() doesn't fill
in the other envelope information.  Making this clean with GetTree
(which is recursive) involves restructuring things so that
scannMessageList() can close the query out, and then make a new query
for the envelope information.

For now, only get the 'From' envelope part.

While here:

- in scanMessage, since we're using the time.Time type anyway, just
  scan directly into it, rather than having a separate variable and
  copying it in.

Signed-off-by: George Dunlap <george.dunlap@cloud.com>