mailfetch: Tweak subject parsing
Add a test for what we want to see pass, and make it pass, by
expanding PatchMailSingleton to accept any text in between PATCH and
the end.

Reorder PatchMailSingleton to the end, so that it doesn't catch more
specific variants.

Add an error message to mailfetch if the "TreeFilterAm" ends up
filtering out all messages.

Signed-off-by: George Dunlap <george.dunlap@cloud.com>